[QAE-456] New objects, CGM pill read func, and extension
Changed bolusEntryTextField so it will support SImple Bolus Calulator and Bolus screens.
Both Current Glucose and Bolus fields use the same identifier so rather than the default value determining the correct field, a supporting extension in Base Screen was added to help distinguish the the two.
This can be used for parsing similar situations in future test cases.